Removed duplicated code
