+++ src/components/layouts/admin-layout.tsx
@@ -1,5 +1,9 @@
+import { Typography } from "@mui/material";
+import { useAppDispatch, useAppSelector } from "app/hooks";
+import GenericSnackbar from "components/generic/generic-snackbar";
 import Header from "components/layout-components/header";
-import React from "react";
+import { handleSnackbarClose, selectSnackbar, setSnackbarOpen } from "features/snackbar-slice";
+import React, { useEffect } from "react";
 import { Content, ContentWrapper } from "styled/layouts/admin-layout";
 import Breadcrumbs from "./breadcrumbs/breadcrumbs";
 import NavigationHeader from "./navigations/navigation-header";
@@ -10,6 +14,18 @@ import NavigationHeader from "./navigations/navigation-header";
  * @param props component properties
  */
 const AdminLayout: React.FC = ({ children }) => {
+  const dispatch = useAppDispatch();
+  const { snackbarMessage, snackbarOpen } = useAppSelector(selectSnackbar);
+
+  /**
+   * Event handler for snackbar open
+   */
+  const handleSnackbarOpen = () => snackbarMessage && dispatch(setSnackbarOpen(true));
+
+  useEffect(() => {
+    handleSnackbarOpen();
+  }, [ snackbarMessage ]);
+
   /**
    * Component render
    */
@@ -19,6 +35,16 @@ const AdminLayout: React.FC = ({ children }) => {
       <NavigationHeader/>
       <Breadcrumbs/>
       <ContentWrapper>
+        <GenericSnackbar
+          open={ snackbarOpen }
+          onClose={ () => dispatch(handleSnackbarClose()) }
+          autoHideDuration={ 4000 }
+          severity="success"
+        >
+          <Typography variant="body2">
+            { snackbarMessage }
+          </Typography>
+        </GenericSnackbar>
         <Content>
           {children}
         </Content>
