+++ src/components/screens/editor/draft-editor-screen.tsx
@@ -15,9 +15,8 @@ import { useApiClient, useAppDispatch, useAppSelector } from "app/hooks";
 import GenericLoaderWrapper from "components/generic/generic-loader";
 import { RoundActionButton } from "styled/generic/form";
 import { selectMetaform, setMetaformVersion } from "features/metaform-slice";
-import { selectSnackbar, setSnackbarMessage, setSnackbarOpen } from "features/snackbar-slice";
+import { setSnackbarMessage } from "features/snackbar-slice";
 import ConfirmDialog from "components/generic/confirm-dialog";
-import GenericSnackbar from "components/generic/generic-snackbar";
 
 /**
  * Draft editor screen component
@@ -33,7 +32,6 @@ const DraftEditorScreen: React.FC = () => {
 
   const dispatch = useAppDispatch();
   const { metaformVersion } = useAppSelector(selectMetaform);
-  const { snackbarMessage, snackbarOpen } = useAppSelector(selectSnackbar);
   
   const draftForm = MetaformUtils.getDraftForm(metaformVersion);
   const editorRef = useRef<HTMLDivElement>(null);
@@ -199,54 +197,25 @@ const DraftEditorScreen: React.FC = () => {
     </Stack>
   );
 
-  /**
-   * Event handler for snackbar close
-   */
-  const handleSnackbarClose = () => {
-    dispatch(setSnackbarOpen(false));
-    dispatch(setSnackbarMessage());
-  };
-
-  /**
-   * Event handler for snackbar open
-   */
-  const handleSnackbarOpen = () => snackbarMessage && dispatch(setSnackbarOpen(true));
-
-  useEffect(() => {
-    handleSnackbarOpen();
-  }, [ snackbarMessage ]);
-
   return (
-    <>
-      <GenericSnackbar
-        open={ snackbarOpen }
-        onClose={ handleSnackbarClose }
-        autoHideDuration={ 4000 }
-        severity="success"
-      >
-        <Typography variant="body2">
-          { snackbarMessage }
-        </Typography>
-      </GenericSnackbar>
-      <Stack flex={ 1 } overflow="hidden">
-        { renderPublishConfirmDialog() }
-        <NavigationTabContainer>
-          <NavigationTab
-            text={ strings.navigationHeader.editorScreens.draftEditorScreen }
-            renderActions={ draftEditorActions }
-          />
-        </NavigationTabContainer>
-        <Divider/>
-        <GenericLoaderWrapper loading={ loading }>
-          <MetaformEditor
-            editorRef={ editorRef }
-            memberGroups={ memberGroups }
-            pendingForm={ MetaformUtils.jsonToMetaform(draftForm) }
-            setPendingForm={ setPendingForm }
-          />
-        </GenericLoaderWrapper>
-      </Stack>
-    </>
+    <Stack flex={ 1 } overflow="hidden">
+      { renderPublishConfirmDialog() }
+      <NavigationTabContainer>
+        <NavigationTab
+          text={ strings.navigationHeader.editorScreens.draftEditorScreen }
+          renderActions={ draftEditorActions }
+        />
+      </NavigationTabContainer>
+      <Divider/>
+      <GenericLoaderWrapper loading={ loading }>
+        <MetaformEditor
+          editorRef={ editorRef }
+          memberGroups={ memberGroups }
+          pendingForm={ MetaformUtils.jsonToMetaform(draftForm) }
+          setPendingForm={ setPendingForm }
+        />
+      </GenericLoaderWrapper>
+    </Stack>
   );
 };
